feat(dedicated): add ipv6 order choice
  ref: MANAGER-13077
