+++ packages/manager/apps/dedicated/client/app/components/ip/agoraOrder/ipv4/ipv4.module.js
@@ -1,0 +1,32 @@
+import angular from 'angular';
+import 'angular-translate';
+import 'oclazyload';
+import 'ovh-api-services';
+import managerCore from '@ovh-ux/manager-core';
+import ngOvhUtils from '@ovh-ux/ng-ovh-utils';
+import '@ovh-ux/ui-kit';
+import '@uirouter/angularjs';
+import service from '../ip-ip-agoraOrder.service';
+
+import routing from './ipv4.routes';
+import controller from './ipv4.controller';
+
+const moduleName = 'ovhManagerDedicatedAgoraOrderIpv4';
+
+angular
+  .module(moduleName, [
+    managerCore,
+    ngOvhUtils,
+    'oc.lazyLoad',
+    'oui',
+    'ovh-api-services',
+    'pascalprecht.translate',
+    'ui.router',
+    'ovhManagerCatalogPrice'
+  ])
+  .config(routing)
+  .controller('AgoraIpV4OrderController', controller)
+  .service('IpAgoraOrder', service)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
